refactor(intune): give the Windows analyzers one masking implementation
Before: `intune::apps::windows::scripts::redaction` owned the UPN, user-profile
and command-line masking regexes outright. A second Windows workload analyzer
needs exactly the same masking, and copying those four regexes would create two
implementations that drift apart -- the class of duplication that already drew
review comment on #386.

This moves the masking primitives to `intune::apps::windows::common`, unchanged.
`scripts::redaction` now calls into it and keeps deciding *what* is sensitive,
which is the part that genuinely differs per analyzer. `scripts::redact_text`
stays exported from the same path it was, so no consumer moves.

This is the right seam because the split is between two different kinds of
decision. Which spans of free text are dangerous is a property of Windows log
text in general. Which fields carry that text is a property of one analyzer's
contract. Only the first is shared.

Verified with `cargo test --locked -p cmtraceopen-parser`: 24
`intune_windows_scripts` integration tests pass unchanged, including the
redaction and golden-export cases, plus 407 unit and 222 esp tests. 0 failed.
`cargo clippy --locked -p cmtraceopen-parser --all-targets -- -D warnings` is
clean.

Refs #360
