+++ crates/cmtraceopen-parser/src/intune/apps/windows/mod.rs
@@ -3,4 +3,5 @@
 //! These modules are pure: they consume artifacts the caller already read and
 //! decoded, and they never touch the filesystem, registry, or network.
 
+pub mod common;
 pub mod scripts;
